+++ src/LuaLib.ts
@@ -2,6 +2,7 @@ import * as path from "path";
 import { EmitHost } from "./transpilation";
 import * as lua from "./LuaAST";
 import { LuaTarget } from "./CompilerOptions";
+import { getOrUpdate } from "./utils";
 
 export enum LuaLibFeature {
     ArrayConcat = "ArrayConcat",
@@ -37,6 +38,7 @@ export enum LuaLibFeature {
     CloneDescriptor = "CloneDescriptor",
     CountVarargs = "CountVarargs",
     Decorate = "Decorate",
+    DecorateLegacy = "DecorateLegacy",
     DecorateParam = "DecorateParam",
     Delete = "Delete",
     DelegatedYield = "DelegatedYield",
@@ -57,6 +59,7 @@ export enum LuaLibFeature {
     NumberIsFinite = "NumberIsFinite",
     NumberIsNaN = "NumberIsNaN",
     NumberToString = "NumberToString",
+    NumberToFixed = "NumberToFixed",
     ObjectAssign = "ObjectAssign",
     ObjectDefineProperty = "ObjectDefineProperty",
     ObjectEntries = "ObjectEntries",
@@ -109,6 +112,7 @@ export interface LuaLibFeatureInfo {
     dependencies?: LuaLibFeature[];
     exports: string[];
 }
+
 export type LuaLibModulesInfo = Record<LuaLibFeature, LuaLibFeatureInfo>;
 
 export function resolveLuaLibDir(luaTarget: LuaTarget) {
@@ -117,27 +121,55 @@ export function resolveLuaLibDir(luaTarget: LuaTarget) {
 }
 
 export const luaLibModulesInfoFileName = "lualib_module_info.json";
-const luaLibModulesInfo = new Map<string, LuaLibModulesInfo>();
+const luaLibModulesInfo = new Map<LuaTarget, LuaLibModulesInfo>();
+
 export function getLuaLibModulesInfo(luaTarget: LuaTarget, emitHost: EmitHost): LuaLibModulesInfo {
-    const lualibPath = path.join(resolveLuaLibDir(luaTarget), luaLibModulesInfoFileName);
-    if (!luaLibModulesInfo.has(lualibPath)) {
+    if (!luaLibModulesInfo.has(luaTarget)) {
+        const lualibPath = path.join(resolveLuaLibDir(luaTarget), luaLibModulesInfoFileName);
         const result = emitHost.readFile(lualibPath);
         if (result !== undefined) {
-            luaLibModulesInfo.set(lualibPath, JSON.parse(result) as LuaLibModulesInfo);
+            luaLibModulesInfo.set(luaTarget, JSON.parse(result) as LuaLibModulesInfo);
         } else {
             throw new Error(`Could not load lualib dependencies from '${lualibPath}'`);
         }
     }
-    return luaLibModulesInfo.get(lualibPath) as LuaLibModulesInfo;
+    return luaLibModulesInfo.get(luaTarget)!;
 }
 
-export function readLuaLibFeature(feature: LuaLibFeature, luaTarget: LuaTarget, emitHost: EmitHost): string {
-    const featurePath = path.join(resolveLuaLibDir(luaTarget), `${feature}.lua`);
-    const luaLibFeature = emitHost.readFile(featurePath);
-    if (luaLibFeature === undefined) {
-        throw new Error(`Could not load lualib feature from '${featurePath}'`);
+// This caches the names of lualib exports to their LuaLibFeature, avoiding a linear search for every lookup
+const lualibExportToFeature = new Map<LuaTarget, ReadonlyMap<string, LuaLibFeature>>();
+
+export function getLuaLibExportToFeatureMap(
+    luaTarget: LuaTarget,
+    emitHost: EmitHost
+): ReadonlyMap<string, LuaLibFeature> {
+    if (!lualibExportToFeature.has(luaTarget)) {
+        const luaLibModulesInfo = getLuaLibModulesInfo(luaTarget, emitHost);
+        const map = new Map<string, LuaLibFeature>();
+        for (const [feature, info] of Object.entries(luaLibModulesInfo)) {
+            for (const exportName of info.exports) {
+                map.set(exportName, feature as LuaLibFeature);
+            }
+        }
+        lualibExportToFeature.set(luaTarget, map);
     }
-    return luaLibFeature;
+
+    return lualibExportToFeature.get(luaTarget)!;
+}
+
+const lualibFeatureCache = new Map<LuaTarget, Map<LuaLibFeature, string>>();
+
+export function readLuaLibFeature(feature: LuaLibFeature, luaTarget: LuaTarget, emitHost: EmitHost): string {
+    const featureMap = getOrUpdate(lualibFeatureCache, luaTarget, () => new Map());
+    if (!featureMap.has(feature)) {
+        const featurePath = path.join(resolveLuaLibDir(luaTarget), `${feature}.lua`);
+        const luaLibFeature = emitHost.readFile(featurePath);
+        if (luaLibFeature === undefined) {
+            throw new Error(`Could not load lualib feature from '${featurePath}'`);
+        }
+        featureMap.set(feature, luaLibFeature);
+    }
+    return featureMap.get(feature)!;
 }
 
 export function resolveRecursiveLualibFeatures(
@@ -173,14 +205,9 @@ export function loadInlineLualibFeatures(
     luaTarget: LuaTarget,
     emitHost: EmitHost
 ): string {
-    let result = "";
-
-    for (const feature of resolveRecursiveLualibFeatures(features, luaTarget, emitHost)) {
-        const luaLibFeature = readLuaLibFeature(feature, luaTarget, emitHost);
-        result += luaLibFeature + "\n";
-    }
-
-    return result;
+    return resolveRecursiveLualibFeatures(features, luaTarget, emitHost)
+        .map(feature => readLuaLibFeature(feature, luaTarget, emitHost))
+        .join("\n");
 }
 
 export function loadImportedLualibFeatures(
@@ -192,13 +219,13 @@ export function loadImportedLualibFeatures(
     const luaLibModuleInfo = getLuaLibModulesInfo(luaTarget, emitHost);
 
     const imports = Array.from(features).flatMap(feature => luaLibModuleInfo[feature].exports);
+    if (imports.length === 0) {
+        return [];
+    }
 
     const requireCall = lua.createCallExpression(lua.createIdentifier("require"), [
         lua.createStringLiteral(luaLibName),
     ]);
-    if (imports.length === 0) {
-        return [];
-    }
 
     const luaLibId = lua.createIdentifier("____lualib");
     const importStatement = lua.createVariableDeclarationStatement(luaLibId, requireCall);
@@ -216,6 +243,7 @@ export function loadImportedLualibFeatures(
 }
 
 const luaLibBundleContent = new Map<string, string>();
+
 export function getLuaLibBundle(luaTarget: LuaTarget, emitHost: EmitHost): string {
     const lualibPath = path.join(resolveLuaLibDir(luaTarget), "lualib_bundle.lua");
     if (!luaLibBundleContent.has(lualibPath)) {
@@ -228,4 +256,43 @@ export function getLuaLibBundle(luaTarget: LuaTarget, emitHost: EmitHost): strin
     }
 
     return luaLibBundleContent.get(lualibPath) as string;
+}
+
+export function getLualibBundleReturn(exportedValues: string[]): string {
+    return `\nreturn {\n${exportedValues.map(exportName => `  ${exportName} = ${exportName}`).join(",\n")}\n}\n`;
+}
+
+export function buildMinimalLualibBundle(
+    features: Iterable<LuaLibFeature>,
+    luaTarget: LuaTarget,
+    emitHost: EmitHost
+): string {
+    const code = loadInlineLualibFeatures(features, luaTarget, emitHost);
+    const moduleInfo = getLuaLibModulesInfo(luaTarget, emitHost);
+    const exports = Array.from(features).flatMap(feature => moduleInfo[feature].exports);
+
+    return code + getLualibBundleReturn(exports);
+}
+
+export function findUsedLualibFeatures(
+    luaTarget: LuaTarget,
+    emitHost: EmitHost,
+    luaContents: string[]
+): Set<LuaLibFeature> {
+    const features = new Set<LuaLibFeature>();
+    const exportToFeatureMap = getLuaLibExportToFeatureMap(luaTarget, emitHost);
+
+    for (const lua of luaContents) {
+        const regex = /^local (\w+) = ____lualib\.(\w+)$/gm;
+        while (true) {
+            const match = regex.exec(lua);
+            if (!match) break;
+            const [, localName, exportName] = match;
+            if (localName !== exportName) continue;
+            const feature = exportToFeatureMap.get(exportName);
+            if (feature) features.add(feature);
+        }
+    }
+
+    return features;
 }