+++ src/LuaPrinter.ts
@@ -236,7 +236,10 @@ export class LuaPrinter {
         const luaTarget = this.options.luaTarget ?? LuaTarget.Universal;
         const luaLibImport = this.options.luaLibImport ?? LuaLibImportKind.Require;
         const luaLibName = this.options.luaLibName ?? "lualib_bundle";
-        if (luaLibImport === LuaLibImportKind.Require && file.luaLibFeatures.size > 0) {
+        if (
+            (luaLibImport === LuaLibImportKind.Require || luaLibImport === LuaLibImportKind.RequireMinimal) &&
+            file.luaLibFeatures.size > 0
+        ) {
             // Import lualib features
             sourceChunks = this.printStatementArray(
                 loadImportedLualibFeatures(file.luaLibFeatures, luaTarget, this.emitHost, luaLibName)
@@ -610,6 +613,8 @@ export class LuaPrinter {
                 return this.printIdentifier(expression as lua.Identifier);
             case lua.SyntaxKind.TableIndexExpression:
                 return this.printTableIndexExpression(expression as lua.TableIndexExpression);
+            case lua.SyntaxKind.ParenthesizedExpression:
+                return this.printParenthesizedExpression(expression as lua.ParenthesizedExpression);
             default:
                 throw new Error(`Tried to print unknown statement kind: ${lua.SyntaxKind[expression.kind]}`);
         }
@@ -818,6 +823,10 @@ export class LuaPrinter {
             chunks.push("[", this.printExpression(expression.index), "]");
         }
         return this.createSourceNode(expression, chunks);
+    }
+
+    public printParenthesizedExpression(expression: lua.ParenthesizedExpression) {
+        return this.createSourceNode(expression, ["(", this.printExpression(expression.expression), ")"]);
     }
 
     public printOperator(kind: lua.Operator): SourceNode {