+++ src/transformation/utils/diagnostics.ts
@@ -1,4 +1,5 @@
 import * as ts from "typescript";
+import * as lua from "../../LuaAST";
 import { LuaTarget, TypeScriptToLuaOptions } from "../../CompilerOptions";
 import { createSerialDiagnosticFactory } from "../../utils";
 import { AnnotationKind } from "./annotations";
@@ -154,4 +155,20 @@ export const unsupportedOptionalCompileMembersOnly = createErrorDiagnosticFactor
 
 export const undefinedInArrayLiteral = createErrorDiagnosticFactory(
     "Array literals may not contain undefined or null."
+);
+
+export const invalidMethodCallExtensionUse = createErrorDiagnosticFactory(
+    "This language extension must be called as a method."
+);
+
+export const invalidSpreadInCallExtension = createErrorDiagnosticFactory(
+    "Spread elements are not supported in call extensions."
+);
+
+export const cannotAssignToNodeOfKind = createErrorDiagnosticFactory(
+    (kind: lua.SyntaxKind) => `Cannot create assignment assigning to a node of type ${lua.SyntaxKind[kind]}.`
+);
+
+export const incompleteFieldDecoratorWarning = createWarningDiagnosticFactory(
+    "You are using a class field decorator, note that tstl ignores returned value initializers!"
 );