+++ test/transpile/module-resolution.spec.ts
@@ -174,6 +174,22 @@ describe("module resolution with sourceDir", () => {
             .setOptions({ luaBundle: "bundle.lua", luaBundleEntry: mainFile })
             .expectToEqual(expectedResult);
     });
+
+    // https://github.com/TypeScriptToLua/TypeScriptToLua/issues/1394
+    test("can resolve files with non-standard extension (#1394)", () => {
+        util.testProject(path.join(projectPath, "tsconfig.json"))
+            .setMainFileName(path.join(projectPath, "src", "main.ts"))
+            .setOptions({ outDir: "tstl-out", extension: ".script" })
+            .expectToEqual(expectedResult);
+    });
+
+    // https://github.com/TypeScriptToLua/TypeScriptToLua/issues/1394
+    test("can resolve files with non-standard extension without separator (#1394)", () => {
+        util.testProject(path.join(projectPath, "tsconfig.json"))
+            .setMainFileName(path.join(projectPath, "src", "main.ts"))
+            .setOptions({ outDir: "tstl-out", extension: "script" })
+            .expectToEqual(expectedResult);
+    });
 });
 
 describe("module resolution project with lua sources", () => {
@@ -268,6 +284,46 @@ describe("module resolution project with dependencies built by tstl library mode
         util.testProject(path.join(projectPath, "tsconfig.json"))
             .setMainFileName(mainFile)
             .setOptions({ luaBundle: "bundle.lua", luaBundleEntry: mainFile })
+            .expectToEqual(expectedResult);
+    });
+});
+
+describe("module resolution project with dependencies built by tstl library mode and has exports field", () => {
+    const projectPath = path.resolve(__dirname, "module-resolution", "project-with-tstl-library-has-exports-field");
+
+    // First compile dependencies into node_modules. NOTE: Actually writing to disk, very slow
+    const dependency1Path = path.join(projectPath, "node_modules", "dependency1");
+    tstl.transpileProject(path.join(dependency1Path, "tsconfig.json"));
+
+    const expectedResult = {
+        dependency1IndexResult: "function in dependency 1 index: dependency1OtherFileFunc in dependency1/d1otherfile",
+        dependency1OtherFileFuncResult: "dependency1OtherFileFunc in dependency1/d1otherfile",
+    };
+
+    test("can resolve lua dependencies", () => {
+        const transpileResult = util
+            .testProject(path.join(projectPath, "tsconfig.json"))
+            .setMainFileName(path.join(projectPath, "main.ts"))
+            .setOptions({ outDir: "tstl-out", moduleResolution: ts.ModuleResolutionKind.Node16 })
+            .expectToEqual(expectedResult)
+            .getLuaResult();
+
+        // Assert node_modules file requires the correct lualib_bundle
+        const requiringLuaFile = path.join("lua_modules", "dependency1", "dist", "index.lua");
+        const lualibRequiringFile = transpileResult.transpiledFiles.find(f => f.outPath.endsWith(requiringLuaFile));
+        expect(lualibRequiringFile).toBeDefined();
+        expect(lualibRequiringFile?.lua).toContain('require("lualib_bundle")');
+    });
+
+    test("can resolve dependencies and bundle", () => {
+        const mainFile = path.join(projectPath, "main.ts");
+        util.testProject(path.join(projectPath, "tsconfig.json"))
+            .setMainFileName(mainFile)
+            .setOptions({
+                luaBundle: "bundle.lua",
+                luaBundleEntry: mainFile,
+                moduleResolution: ts.ModuleResolutionKind.Node16,
+            })
             .expectToEqual(expectedResult);
     });
 });